requirements.txt
moviepy==1.0.3
openai==1.12.0
python-dotenv==0.21.0
streamlit==1.31.1